fixed a lot of minor issues and refactored the displayPlaces function into separate file
