+++ index.html
@@ -95,7 +95,8 @@
         </nav>
 
     </footer>
-    <script type="text/javascript" src="./scripts/home.js"></script>
+    <script type="module" src="./scripts/home.js"></script>
+    <script type="module" src="./scripts/placesFunctions.js"></script>
 </body>
 
 </html>
